Revert "Made base header component, updated routing/component passthrough"
This reverts commit 3c5a35e4985f8daca5e6aed0471bf4a7afa17e85.
